Refactor authentication management: Rename AuthManager to PacAuthManager and update imports
