+++ src/client/power-pages/actions-hub/Constants.ts
@@ -34,21 +34,3 @@ export const Constants = {
     }
 };
 
-export interface AuthInfo {
-    userType: string;
-    cloud: string;
-    tenantId: string;
-    tenantCountry: string;
-    user: string;
-    entraIdObjectId: string;
-    puid: string;
-    userCountryRegion: string;
-    tokenExpires: string;
-    authority: string;
-    environmentGeo: string;
-    environmentId: string;
-    environmentType: string;
-    organizationId: string;
-    organizationUniqueName: string;
-    organizationFriendlyName: string;
-}
